server/requirements.txt
# Core OpenEnv dependencies
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0

# Baseline agent (optional — only needed when running agents/)
openai>=1.0.0
